% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function 5.R
\name{average_reign_length_by_gender}
\alias{average_reign_length_by_gender}
\title{Calculate Average Reign Length by Gender}
\usage{
average_reign_length_by_gender(data = wrestling_data_clean)
}
\arguments{
\item{data}{A data frame containing wrestling data. Defaults to \code{wrestling_data_clean}.}
}
\value{
A data frame with each gender and the average reign length for that gender, sorted in descending order.
}
\description{
This function calculates the average championship reign length for each gender.
}
\examples{
average_reign_length_by_gender()
}
