% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_shiny_app.R
\name{launch_shiny_app}
\alias{launch_shiny_app}
\title{Launch the Shiny App from the inst Folder}
\usage{
launch_shiny_app(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Logical; if TRUE, opens the app in the default web browser.}
}
\value{
No return value; the function is called for its side effect of launching the app.
}
\description{
This function launches a Shiny application located in the inst folder of the package.
}
\examples{
launch_shiny_app()
}
