% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function 3.R
\name{top_wrestlers_by_reign}
\alias{top_wrestlers_by_reign}
\title{Get Top Wrestlers by Average Reign Length}
\usage{
top_wrestlers_by_reign(data = wrestling_data_clean, n = 10)
}
\arguments{
\item{data}{A data frame containing wrestling data. Defaults to \code{wrestling_data_clean}.}

\item{n}{The number of top wrestlers to return. Defaults to 10.}
}
\value{
A data frame with the top \code{n} wrestlers and their average reign length.
}
\description{
This function retrieves the top wrestlers based on the average length of their title reigns.
}
\examples{
top_wrestlers_by_reign(n = 5)
}
